+++ common/src/main/java/io/github/mortuusars/exposure/mixin/HideCrosshairMixin.java
@@ -14,18 +14,18 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(Gui.class)
 public abstract class HideCrosshairMixin {
-    @Inject(method = "renderCrosshair", at = @At(value = "HEAD"), cancellable = true)
-    private void renderCrosshair(GuiGraphics guiGraphics, CallbackInfo ci) {
-        Minecraft mc = Minecraft.getInstance();
-        if (Config.Client.PHOTOGRAPH_IN_HAND_HIDE_CROSSHAIR() && mc.player != null && mc.player.getXRot() > 25f
-                && (mc.player.getMainHandItem().getItem() instanceof PhotographItem || mc.player.getMainHandItem().getItem() instanceof StackedPhotographsItem)
-                && mc.player.getOffhandItem().isEmpty())
+    @Inject(method = "render", at = @At(value = "HEAD"), cancellable = true)
+    private void renderGui(GuiGraphics guiGraphics, float partialTick, CallbackInfo ci) {
+        if (ViewfinderClient.isLookingThrough())
             ci.cancel();
     }
 
-    @Inject(method = "render", at = @At(value = "HEAD"), cancellable = true)
-    private void renderCrosshair(GuiGraphics guiGraphics, float partialTick, CallbackInfo ci) {
-        if (ViewfinderClient.isLookingThrough())
+    @Inject(method = "renderCrosshair", at = @At(value = "HEAD"), cancellable = true)
+    private void renderCrosshair(GuiGraphics guiGraphics, CallbackInfo ci) {
+        Minecraft mc = Minecraft.getInstance();
+        if (Config.Client.PHOTOGRAPH_IN_HAND_HIDE_CROSSHAIR.get() && mc.player != null && mc.player.getXRot() > 25f
+                && (mc.player.getMainHandItem().getItem() instanceof PhotographItem || mc.player.getMainHandItem().getItem() instanceof StackedPhotographsItem)
+                && mc.player.getOffhandItem().isEmpty())
             ci.cancel();
     }
 }