+++ fabric/src/main/java/io/github/mortuusars/exposure/fabric/mixin/ItemFrameRendererFabricMixin.java
@@ -20,8 +20,18 @@ public abstract class ItemFrameRendererFabricMixin<T extends ItemFrame>
     }
 
     @Inject(method = "render(Lnet/minecraft/world/entity/decoration/ItemFrame;FFLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;I)V",
-            at = @At(value = "INVOKE", target = "Lcom/mojang/blaze3d/vertex/PoseStack;popPose()V", ordinal = 1))
+            cancellable = true,
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/entity/ItemRenderer;renderStatic(Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/world/item/ItemDisplayContext;IILcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;Lnet/minecraft/world/level/Level;I)V"))
     void onItemFrameRender(T entity, float entityYaw, float partialTicks, PoseStack poseStack, MultiBufferSource buffer, int packedLight, CallbackInfo ci) {
-        ItemFramePhotographRenderer.render(entity, poseStack, buffer, packedLight);
+        poseStack.scale(2F, 2F, 2F);
+        boolean rendered = ItemFramePhotographRenderer.render(entity, poseStack, buffer, packedLight);
+
+        if (rendered) {
+            poseStack.popPose();
+            ci.cancel();
+        }
+        else {
+            poseStack.scale(0.5F, 0.5F, 0.5F);
+        }
     }
 }